requirements.txt
appdirs==1.4.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Login==0.5.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
pipenv==2020.11.15
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
rsa==4.6
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.55.1
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
Flask-Admin==1.5.7
SQLAlchemy==1.4.1
Flask-SQLAlchemy==2.5.1
Markdown